simple_wavetable_synth_host_test/pyproject.toml
[tool.poetry]
name = "simple-wavetable-synth-host-test"
version = "0.1.0"
description = ""
authors = ["Andres V <andres.vazquezbrennan@physics.ox.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.1"
matplotlib = "^3.9.1.post1"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
